Refactor label drawing into LabelRenderSystem
