+++ viewer3d/viewer3dcontroller.cpp
@@ -56,9 +56,9 @@
 #include "render/scenerenderer.h"
 #include "render/render_pipeline.h"
 #include "culling/visibilitysystem.h"
+#include "labels/label_render_system.h"
 #include "picking/selectionsystem.h"
 
-#include <wx/tokenzr.h>
 #include <wx/wx.h>
 #define NANOVG_GL2_IMPLEMENTATION
 #include <algorithm>
@@ -67,7 +67,6 @@
 #include <cfloat>
 #include <cmath>
 #include <filesystem>
-#include <iomanip>
 #include <iostream>
 #include <nanovg.h>
 #include <nanovg_gl.h>
@@ -80,23 +79,10 @@
 
 namespace fs = std::filesystem;
 
-// Font size for on-screen labels drawn with NanoVG in the 3D viewer
-static constexpr float LABEL_FONT_SIZE_3D = 18.0f;
-// Maximum width for on-screen labels before wrapping
-static constexpr float LABEL_MAX_WIDTH = 300.0f;
-// Width of fixture labels in meters for the 2D view
-// Pixels per meter used by the 2D view
-static constexpr float PIXELS_PER_METER = 25.0f;
-
 struct LineRenderProfile {
   float lineWidth = 1.0f;
   bool enableLineSmoothing = false;
 };
-
-static bool ShouldTraceLabelOrder() {
-  static const bool enabled = std::getenv("PERASTAGE_TRACE_LABELS") != nullptr;
-  return enabled;
-}
 
 static std::unordered_set<std::string>
 SnapshotHiddenLayers(const ConfigManager &cfg) {
@@ -365,16 +351,6 @@ static SymbolBounds ComputeSymbolBounds(const CommandBuffer &buffer) {
   return bounds;
 }
 
-static std::string FormatMeters(float mm) {
-  std::ostringstream oss;
-  oss << std::fixed << std::setprecision(2) << mm / 1000.0f;
-  std::string s = oss.str();
-  s.erase(s.find_last_not_of('0') + 1, std::string::npos);
-  if (!s.empty() && s.back() == '.')
-    s.pop_back();
-  return s;
-}
-
 static bool HexToRGB(const std::string &hex, float &r, float &g, float &b) {
   if (hex.size() != 7 || hex[0] != '#')
     return false;
@@ -408,14 +384,6 @@ static void MatrixToArray(const Matrix &m, float out[16]) {
   out[15] = 1.0f;
 }
 
-struct ScreenRect {
-  double minX = DBL_MAX;
-  double minY = DBL_MAX;
-  double maxX = -DBL_MAX;
-  double maxY = -DBL_MAX;
-};
-
-
 struct CullingSettings {
   bool enabled = true;
   float minPixels3D = 2.0f;
@@ -428,230 +396,6 @@ static CullingSettings GetCullingSettings3D(const ConfigManager &cfg) {
   s.minPixels3D = std::max(0.0f, cfg.GetFloat("render_culling_min_pixels_3d"));
   s.minPixels2D = std::max(0.0f, cfg.GetFloat("render_culling_min_pixels_2d"));
   return s;
-}
-
-static int GetLabelLimit(const ConfigManager &cfg, const char *key) {
-  return std::max(0, static_cast<int>(std::lround(cfg.GetFloat(key))));
-}
-
-static bool ProjectBoundingBoxToScreen(const std::array<float, 3> &bbMin,
-                                       const std::array<float, 3> &bbMax,
-                                       int viewportHeight,
-                                       const double model[16],
-                                       const double proj[16],
-                                       const int viewport[4],
-                                       ScreenRect &outRect,
-                                       bool &outAnyDepthVisible) {
-  outRect = ScreenRect{};
-  outAnyDepthVisible = false;
-  bool projected = false;
-
-  std::array<std::array<float, 3>, 8> corners = {
-      std::array<float, 3>{bbMin[0], bbMin[1], bbMin[2]},
-      {bbMax[0], bbMin[1], bbMin[2]},
-      {bbMin[0], bbMax[1], bbMin[2]},
-      {bbMax[0], bbMax[1], bbMin[2]},
-      {bbMin[0], bbMin[1], bbMax[2]},
-      {bbMax[0], bbMin[1], bbMax[2]},
-      {bbMin[0], bbMax[1], bbMax[2]},
-      {bbMax[0], bbMax[1], bbMax[2]}};
-
-  for (const auto &c : corners) {
-    double sx, sy, sz;
-    if (gluProject(c[0], c[1], c[2], model, proj, viewport, &sx, &sy, &sz) ==
-        GL_TRUE) {
-      projected = true;
-      outRect.minX = std::min(outRect.minX, sx);
-      outRect.maxX = std::max(outRect.maxX, sx);
-      const double sy2 = static_cast<double>(viewportHeight) - sy;
-      outRect.minY = std::min(outRect.minY, sy2);
-      outRect.maxY = std::max(outRect.maxY, sy2);
-      if (sz >= 0.0 && sz <= 1.0)
-        outAnyDepthVisible = true;
-    }
-  }
-
-  return projected;
-}
-
-static bool ShouldCullByScreenRect(const ScreenRect &rect, int width, int height,
-                                   float minPixels) {
-  if (rect.maxX < 0.0 || rect.minX > static_cast<double>(width) ||
-      rect.maxY < 0.0 || rect.minY > static_cast<double>(height)) {
-    return true;
-  }
-
-  const double screenWidth = rect.maxX - rect.minX;
-  const double screenHeight = rect.maxY - rect.minY;
-  return screenWidth < static_cast<double>(minPixels) &&
-         screenHeight < static_cast<double>(minPixels);
-}
-
-// Inserts a line break every two words in the provided text.
-static wxString WrapEveryTwoWords(const wxString &text) {
-  wxStringTokenizer tk(text, " ");
-  wxString result;
-  int count = 0;
-  while (tk.HasMoreTokens()) {
-    if (count > 0) {
-      if (count % 2 == 0)
-        result += "\n";
-      else
-        result += " ";
-    }
-    result += tk.GetNextToken();
-    ++count;
-  }
-  return result;
-}
-
-// Draws a text string at screen coordinates using NanoVG. The font size and
-// maximum width are specified in pixels.
-static void DrawText2D(NVGcontext *vg, int font, const std::string &text, int x,
-                       int y, float fontSize = LABEL_FONT_SIZE_3D,
-                       float maxWidth = LABEL_MAX_WIDTH,
-                       bool drawBackground = true, bool drawBorder = true,
-                       NVGcolor textColor = nvgRGBAf(1.f, 1.f, 1.f, 1.f)) {
-  if (!vg || font < 0 || text.empty())
-    return;
-
-  std::string normalizedText = text;
-  size_t pos = 0;
-  while ((pos = normalizedText.find("\r\n", pos)) != std::string::npos) {
-    normalizedText.replace(pos, 2, "\n");
-  }
-  for (char &ch : normalizedText) {
-    if (ch == '\r')
-      ch = '\n';
-  }
-
-  GLint vp[4];
-  glGetIntegerv(GL_VIEWPORT, vp);
-
-  nvgBeginFrame(vg, vp[2], vp[3], 1.0f);
-  nvgSave(vg);
-  nvgFontSize(vg, fontSize);
-  nvgFontFaceId(vg, font);
-  // Center text for multiline labels
-  nvgTextAlign(vg, NVG_ALIGN_CENTER | NVG_ALIGN_MIDDLE);
-
-  // Determine the width based on the actual text content so the background
-  // box tightly fits the rendered label. We measure each line separately to
-  // avoid the fixed width imposed by nvgTextBoxBounds and clamp to a maximum
-  // so long names wrap instead of growing indefinitely.
-
-  float textWidth = 0.0f;
-  size_t start = 0;
-  while (start <= normalizedText.size()) {
-    size_t end = normalizedText.find('\n', start);
-    std::string line = normalizedText.substr(start, end - start);
-    float lb[4];
-    nvgTextBounds(vg, 0.f, 0.f, line.c_str(), nullptr, lb);
-    textWidth = std::max(textWidth, lb[2] - lb[0]);
-    if (end == std::string::npos)
-      break;
-    start = end + 1;
-  }
-  if (maxWidth > 0.0f)
-    textWidth = std::min(textWidth, maxWidth);
-  const int padding = 4;
-
-  // Calculate the exact bounding box for the text using the same alignment
-  // and width that will be used when rendering it. This ensures the
-  // background rectangle matches the visual position of the text.
-  float bounds[4];
-  nvgTextBoxBounds(vg, (float)x, (float)y, textWidth,
-                   normalizedText.c_str(), nullptr, bounds);
-
-  if (drawBackground) {
-    nvgBeginPath(vg);
-    nvgRect(vg, bounds[0] - padding, bounds[1] - padding,
-            (bounds[2] - bounds[0]) + padding * 2,
-            (bounds[3] - bounds[1]) + padding * 2);
-    nvgFillColor(vg, nvgRGBAf(0.f, 0.f, 0.f, 0.6f));
-    nvgFill(vg);
-  }
-
-  if (drawBorder) {
-    nvgBeginPath(vg);
-    nvgRect(vg, bounds[0] - padding, bounds[1] - padding,
-            (bounds[2] - bounds[0]) + padding * 2,
-            (bounds[3] - bounds[1]) + padding * 2);
-    nvgStrokeColor(vg, nvgRGBAf(1.f, 1.f, 1.f, 0.8f));
-    nvgStrokeWidth(vg, 1.0f);
-    nvgStroke(vg);
-  }
-
-  nvgFillColor(vg, textColor);
-  // Draw multi-line label using textWidth to avoid excessive empty space.
-  nvgTextBox(vg, (float)x, (float)y, textWidth, normalizedText.c_str(),
-             nullptr);
-  nvgRestore(vg);
-  nvgEndFrame(vg);
-}
-
-struct LabelLine2D {
-  int font;
-  std::string text;
-  float size;
-  std::string fontFamily;
-};
-
-static void
-DrawLabelLines2D(NVGcontext *vg, const std::vector<LabelLine2D> &lines, int x,
-                 int y, NVGcolor textColor = nvgRGBAf(1.f, 1.f, 1.f, 1.f),
-                 NVGcolor outlineColor = nvgRGBAf(0.f, 0.f, 0.f, 1.f),
-                 bool outline = false) {
-  if (!vg || lines.empty())
-    return;
-
-  GLint vp[4];
-  glGetIntegerv(GL_VIEWPORT, vp);
-  nvgBeginFrame(vg, vp[2], vp[3], 1.0f);
-  nvgSave(vg);
-
-  const float lineSpacing = 2.0f;
-  std::vector<float> heights(lines.size());
-  for (size_t i = 0; i < lines.size(); ++i) {
-    nvgFontSize(vg, lines[i].size);
-    nvgFontFaceId(vg, lines[i].font);
-    nvgTextAlign(vg, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
-    float bounds[4];
-    nvgTextBounds(vg, 0.f, 0.f, lines[i].text.c_str(), nullptr, bounds);
-    heights[i] = bounds[3] - bounds[1];
-  }
-
-  float totalHeight = 0.0f;
-  for (size_t i = 0; i < heights.size(); ++i) {
-    totalHeight += heights[i];
-    if (i + 1 < heights.size())
-      totalHeight += lineSpacing;
-  }
-
-  float currentY = y - totalHeight * 0.5f;
-  for (size_t i = 0; i < lines.size(); ++i) {
-    nvgFontSize(vg, lines[i].size);
-    nvgFontFaceId(vg, lines[i].font);
-    nvgTextAlign(vg, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
-    if (outline) {
-      nvgFillColor(vg, outlineColor);
-      const std::array<std::array<float, 2>, 8> offsets = {
-          std::array<float, 2>{-1.f, 0.f}, std::array<float, 2>{1.f, 0.f},
-          std::array<float, 2>{0.f, -1.f}, std::array<float, 2>{0.f, 1.f},
-          std::array<float, 2>{-1.f, -1.f}, std::array<float, 2>{1.f, -1.f},
-          std::array<float, 2>{-1.f, 1.f}, std::array<float, 2>{1.f, 1.f}};
-      for (const auto &offset : offsets) {
-        nvgText(vg, static_cast<float>(x) + offset[0], currentY + offset[1],
-                lines[i].text.c_str(), nullptr);
-      }
-    }
-    nvgFillColor(vg, textColor);
-    nvgText(vg, (float)x, currentY, lines[i].text.c_str(), nullptr);
-    currentY += heights[i] + lineSpacing;
-  }
-
-  nvgRestore(vg);
-  nvgEndFrame(vg);
 }
 
 static std::array<float, 3> TransformPoint(const Matrix &m,
@@ -790,7 +534,8 @@ void Viewer3DController::RecordText(float x, float y, const std::string &text,
 Viewer3DController::Viewer3DController()
     : m_sceneRenderer(std::make_unique<SceneRenderer>(*this)),
       m_visibilitySystem(std::make_unique<VisibilitySystem>(*this)),
-      m_selectionSystem(std::make_unique<SelectionSystem>(*this)) {
+      m_selectionSystem(std::make_unique<SelectionSystem>(*this)),
+      m_labelRenderSystem(std::make_unique<LabelRenderSystem>(*this)) {
   // Actual initialization of OpenGL-dependent resources is delayed
   // until a valid context is available.
 }
@@ -2820,98 +2565,7 @@ void Viewer3DController::SetupBasicLighting() {
 }
 
 void Viewer3DController::DrawFixtureLabels(int width, int height) {
-  ConfigManager &cfg = ConfigManager::Get();
-  double model[16];
-  double proj[16];
-  int viewport[4];
-  glGetDoublev(GL_MODELVIEW_MATRIX, model);
-  glGetDoublev(GL_PROJECTION_MATRIX, proj);
-  glGetIntegerv(GL_VIEWPORT, viewport);
-
-  const auto hiddenLayers = SnapshotHiddenLayers(cfg);
-  const CullingSettings culling = GetCullingSettings3D(cfg);
-  const float minLabelPixels = culling.minPixels3D;
-  const bool useLabelOptimizations =
-      cfg.GetFloat("label_optimizations_enabled") >= 0.5f;
-  bool showName = cfg.GetFloat("label_show_name") != 0.0f;
-  bool showId = cfg.GetFloat("label_show_id") != 0.0f;
-  bool showDmx = cfg.GetFloat("label_show_dmx") != 0.0f;
-  // The 3D view does not scale labels with any zoom factor, but the
-  // configuration values are multiplied by a zoom parameter in the 2D view.
-  // Use a fixed zoom of 1.0f here so that the code compiles and label sizes
-  // are consistent with the configured defaults.
-  float zoom = 1.0f;
-  float nameSize = cfg.GetFloat("label_font_size_name") * zoom;
-  float idSize = cfg.GetFloat("label_font_size_id") * zoom;
-  float dmxSize = cfg.GetFloat("label_font_size_dmx") * zoom;
-
-  const auto &fixtures = SceneDataManager::Instance().GetFixtures();
-  ViewFrustumSnapshot frustum{};
-  std::copy(std::begin(viewport), std::end(viewport), std::begin(frustum.viewport));
-  std::copy(std::begin(model), std::end(model), std::begin(frustum.model));
-  std::copy(std::begin(proj), std::end(proj), std::begin(frustum.projection));
-  const VisibleSet &visibleSet =
-      GetVisibleSet(frustum, hiddenLayers, culling.enabled, minLabelPixels);
-  for (const auto &uuid : visibleSet.fixtureUuids) {
-    auto fixtureIt = fixtures.find(uuid);
-    if (fixtureIt == fixtures.end())
-      continue;
-    const auto &f = fixtureIt->second;
-    if (uuid != m_highlightUuid)
-      continue;
-    double sx, sy, sz;
-    // Use bounding box center if available
-    auto bit = m_fixtureBounds.find(uuid);
-    if (useLabelOptimizations && culling.enabled && bit != m_fixtureBounds.end()) {
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max, height,
-                                      model, proj, viewport, rect,
-                                      anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, width, height, minLabelPixels)) {
-        continue;
-      }
-    }
-    if (bit != m_fixtureBounds.end()) {
-      const BoundingBox &bb = bit->second;
-      double wx = (bb.min[0] + bb.max[0]) * 0.5;
-      double wy = (bb.min[1] + bb.max[1]) * 0.5;
-      double wz = (bb.min[2] + bb.max[2]) * 0.5;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    } else {
-      double wx = f.transform.o[0] * RENDER_SCALE;
-      double wy = f.transform.o[1] * RENDER_SCALE;
-      double wz = f.transform.o[2] * RENDER_SCALE;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    }
-
-    int x = static_cast<int>(sx);
-    int y = height - static_cast<int>(sy);
-    wxString label;
-    if (showName)
-      label = f.instanceName.empty() ? wxString::FromUTF8(uuid)
-                                     : wxString::FromUTF8(f.instanceName);
-    if (showId) {
-      if (!label.empty())
-        label += "\n";
-      label += "ID: " + wxString::Format("%d", f.fixtureId);
-    }
-    if (showDmx && !f.address.empty()) {
-      if (!label.empty())
-        label += "\n";
-      label += wxString::FromUTF8(f.address);
-    }
-    if (label.empty())
-      continue;
-
-    auto utf8 = label.ToUTF8();
-    DrawText2D(m_vg, m_font, std::string(utf8.data(), utf8.length()), x, y);
-  }
+  m_labelRenderSystem->DrawFixtureLabels(width, height);
 }
 
 // Renders labels for all fixtures in the current scene. Each label displays the
@@ -2923,456 +2577,15 @@ void Viewer3DController::DrawFixtureLabels(int width, int height) {
 // scales the label like regular geometry when zooming the 2D view.
 void Viewer3DController::DrawAllFixtureLabels(int width, int height,
                                               Viewer2DView view, float zoom) {
-  ConfigManager &cfg = ConfigManager::Get();
-  double model[16];
-  double proj[16];
-  int viewport[4];
-  glGetDoublev(GL_MODELVIEW_MATRIX, model);
-  glGetDoublev(GL_PROJECTION_MATRIX, proj);
-  glGetIntegerv(GL_VIEWPORT, viewport);
-
-  const auto hiddenLayers = SnapshotHiddenLayers(cfg);
-  const std::array<const char *, 4> nameKeys = {"label_show_name_top",
-                                               "label_show_name_front",
-                                               "label_show_name_side",
-                                               "label_show_name_top"};
-  const std::array<const char *, 4> idKeys = {"label_show_id_top",
-                                             "label_show_id_front",
-                                             "label_show_id_side",
-                                             "label_show_id_top"};
-  const std::array<const char *, 4> dmxKeys = {"label_show_dmx_top",
-                                              "label_show_dmx_front",
-                                              "label_show_dmx_side",
-                                              "label_show_dmx_top"};
-  const std::array<const char *, 4> distKeys = {"label_offset_distance_top",
-                                                "label_offset_distance_front",
-                                                "label_offset_distance_side",
-                                                "label_offset_distance_top"};
-  const std::array<const char *, 4> angleKeys = {"label_offset_angle_top",
-                                                 "label_offset_angle_front",
-                                                 "label_offset_angle_side",
-                                                 "label_offset_angle_top"};
-  int viewIdx = static_cast<int>(view);
-  bool showName = cfg.GetFloat(nameKeys[viewIdx]) != 0.0f;
-  bool showId = cfg.GetFloat(idKeys[viewIdx]) != 0.0f;
-  bool showDmx = cfg.GetFloat(dmxKeys[viewIdx]) != 0.0f;
-  float nameSize = cfg.GetFloat("label_font_size_name") * zoom;
-  float idSize = cfg.GetFloat("label_font_size_id") * zoom;
-  float dmxSize = cfg.GetFloat("label_font_size_dmx") * zoom;
-  float labelDist = cfg.GetFloat(distKeys[viewIdx]);
-  float labelAngle = cfg.GetFloat(angleKeys[viewIdx]);
-  constexpr float deg2rad = 3.14159265358979323846f / 180.0f;
-  float angRad = labelAngle * deg2rad;
-  float offX = 0.0f;
-  float offY = 0.0f;
-  float offZ = 0.0f;
-  switch (static_cast<Viewer2DView>(viewIdx)) {
-  case Viewer2DView::Top:
-    offX = labelDist * std::sin(angRad);
-    offY = labelDist * std::cos(angRad);
-    break;
-  case Viewer2DView::Bottom:
-    offX = labelDist * std::sin(angRad);
-    offY = labelDist * std::cos(angRad);
-    break;
-  case Viewer2DView::Front:
-    offX = labelDist * std::sin(angRad);
-    offZ = labelDist * std::cos(angRad);
-    break;
-  case Viewer2DView::Side:
-    offY = -labelDist * std::sin(angRad);
-    offZ = labelDist * std::cos(angRad);
-    break;
-  }
-
-  const CullingSettings culling = GetCullingSettings3D(cfg);
-  const float minLabelPixels = culling.minPixels2D;
-  const bool useLabelOptimizations =
-      cfg.GetFloat("label_optimizations_enabled") >= 0.5f;
-  const int maxFixtureLabels = GetLabelLimit(cfg, "label_max_fixtures");
-
-  struct FixtureLabelCandidate {
-    const std::string *uuid = nullptr;
-    const Fixture *fixture = nullptr;
-    double area = 0.0;
-  };
-  std::vector<FixtureLabelCandidate> candidates;
-
-  const auto &fixtures = SceneDataManager::Instance().GetFixtures();
-  candidates.reserve(fixtures.size());
-  for (const auto &[uuid, f] : fixtures) {
-    if (!IsLayerVisibleCached(hiddenLayers, f.layer))
-      continue;
-
-    auto bit = m_fixtureBounds.find(uuid);
-    if (useLabelOptimizations && culling.enabled && bit != m_fixtureBounds.end()) {
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max, height,
-                                      model, proj, viewport, rect,
-                                      anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, width, height, minLabelPixels)) {
-        continue;
-      }
-      const double area = std::max(0.0, rect.maxX - rect.minX) *
-                          std::max(0.0, rect.maxY - rect.minY);
-      candidates.push_back({&uuid, &f, area});
-    } else {
-      candidates.push_back({&uuid, &f, 0.0});
-    }
-  }
-
-  if (useLabelOptimizations && maxFixtureLabels > 0 &&
-      static_cast<int>(candidates.size()) > maxFixtureLabels) {
-    std::partial_sort(candidates.begin(),
-                      candidates.begin() + maxFixtureLabels,
-                      candidates.end(), [](const auto &a, const auto &b) {
-                        return a.area > b.area;
-                      });
-    candidates.resize(maxFixtureLabels);
-  }
-
-  for (const auto &candidate : candidates) {
-    const std::string &uuid = *candidate.uuid;
-    const Fixture &f = *candidate.fixture;
-
-    double wx, wy, wz;
-    auto bit = m_fixtureBounds.find(uuid);
-    if (bit != m_fixtureBounds.end()) {
-      const BoundingBox &bb = bit->second;
-      double cx = (bb.min[0] + bb.max[0]) * 0.5;
-      double cy = (bb.min[1] + bb.max[1]) * 0.5;
-      double cz = (bb.min[2] + bb.max[2]) * 0.5;
-
-      // Anchor label offsets from the top of the fixture relative to the active
-      // view so the distance matches the PDF output expectation. Top view uses
-      // +Y as "up" while front/side views share +Z.
-      switch (static_cast<Viewer2DView>(viewIdx)) {
-      case Viewer2DView::Top:
-        cy = bb.max[1];
-        break;
-      case Viewer2DView::Bottom:
-        cy = bb.max[1];
-        break;
-      case Viewer2DView::Front:
-      case Viewer2DView::Side:
-        cz = bb.max[2];
-        break;
-      }
-
-      wx = cx + offX;
-      wy = cy + offY;
-      wz = cz + offZ;
-    } else {
-      double cx = f.transform.o[0] * RENDER_SCALE;
-      double cy = f.transform.o[1] * RENDER_SCALE;
-      double cz = f.transform.o[2] * RENDER_SCALE;
-      wx = cx + offX;
-      wy = cy + offY;
-      wz = cz + offZ;
-    }
-
-    double sx, sy, sz;
-    if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) != GL_TRUE)
-      continue;
-
-    int x = static_cast<int>(sx);
-    // Convert OpenGL's origin to top-left.
-    int y = height - static_cast<int>(sy);
-
-    constexpr const char *kRegularFamily = "sans";
-    constexpr const char *kBoldFamily = "sans-bold";
-    std::vector<LabelLine2D> lines;
-    if (showName) {
-      wxString baseName = f.instanceName.empty()
-                              ? wxString::FromUTF8(uuid)
-                              : wxString::FromUTF8(f.instanceName);
-      wxString wrapped = WrapEveryTwoWords(baseName);
-      wxStringTokenizer nameLines(wrapped, "\n");
-      while (nameLines.HasMoreTokens()) {
-        wxString line = nameLines.GetNextToken();
-        auto utf8 = line.ToUTF8();
-        lines.push_back({m_font, std::string(utf8.data(), utf8.length()),
-                         nameSize, kRegularFamily});
-      }
-    }
-    if (showId) {
-      wxString idLine = "ID: " + wxString::Format("%d", f.fixtureId);
-      auto utf8 = idLine.ToUTF8();
-      lines.push_back({m_font, std::string(utf8.data(), utf8.length()), idSize,
-                       kRegularFamily});
-    }
-    if (showDmx && !f.address.empty()) {
-      int dmxFont = m_fontBold >= 0 ? m_fontBold : m_font;
-      wxString addrLine = wxString::FromUTF8(f.address);
-      auto utf8 = addrLine.ToUTF8();
-      lines.push_back({dmxFont, std::string(utf8.data(), utf8.length()),
-                       dmxSize, kBoldFamily});
-    }
-    if (lines.empty())
-      continue;
-
-    if (m_captureCanvas) {
-      std::string labelSourceKey = "label:" + uuid;
-      m_captureCanvas->SetSourceKey(labelSourceKey);
-
-      const float pxToWorld = 1.0f / (PIXELS_PER_METER * zoom);
-      const float lineSpacingWorld = 2.0f * pxToWorld;
-
-      std::vector<float> worldFontSizes;
-      std::vector<float> lineHeightsWorld;
-      std::vector<float> ascentsWorld;
-      std::vector<float> descentsWorld;
-      worldFontSizes.reserve(lines.size());
-      lineHeightsWorld.reserve(lines.size());
-      ascentsWorld.reserve(lines.size());
-      descentsWorld.reserve(lines.size());
-
-      // Measure each line with NanoVG so the PDF layout matches the on-screen
-      // rendering. Measurements happen in pixel space and are converted to the
-      // world-space units consumed by the PDF exporter.
-      for (const auto &ln : lines) {
-        worldFontSizes.push_back(ln.size * pxToWorld);
-        nvgFontSize(m_vg, ln.size);
-        nvgFontFaceId(m_vg, ln.font);
-        nvgTextAlign(m_vg, NVG_ALIGN_CENTER | NVG_ALIGN_TOP);
-        float bounds[4];
-        nvgTextBounds(m_vg, 0.f, 0.f, ln.text.c_str(), nullptr, bounds);
-        lineHeightsWorld.push_back((bounds[3] - bounds[1]) * pxToWorld);
-        float ascender = 0.0f;
-        float descender = 0.0f;
-        float lineh = 0.0f;
-        nvgTextMetrics(m_vg, &ascender, &descender, &lineh);
-        ascentsWorld.push_back(ascender * pxToWorld);
-        descentsWorld.push_back(-descender * pxToWorld);
-      }
-
-      float totalHeight = 0.0f;
-      for (size_t i = 0; i < lineHeightsWorld.size(); ++i) {
-        totalHeight += lineHeightsWorld[i];
-        if (i + 1 < lineHeightsWorld.size())
-          totalHeight += lineSpacingWorld;
-      }
-
-      auto toPlan2D = [](double wx, double wy, double wz, Viewer2DView view) {
-        switch (view) {
-        case Viewer2DView::Top:
-          return std::array<float, 2>{static_cast<float>(wx),
-                                      static_cast<float>(wy)};
-        case Viewer2DView::Bottom:
-          return std::array<float, 2>{static_cast<float>(wx),
-                                      static_cast<float>(wy)};
-        case Viewer2DView::Front:
-          return std::array<float, 2>{static_cast<float>(wx),
-                                      static_cast<float>(wz)};
-        case Viewer2DView::Side:
-          return std::array<float, 2>{static_cast<float>(-wy),
-                                      static_cast<float>(wz)};
-        }
-        return std::array<float, 2>{static_cast<float>(wx),
-                                    static_cast<float>(wy)};
-      };
-
-      auto anchor =
-          toPlan2D(wx, wy, wz, static_cast<Viewer2DView>(viewIdx));
-      // Start from the top of the label block instead of the bottom.
-      // We subtract the ascent from currentY so the baseline matches the
-      // top-aligned position.
-      // After recording each line, we move downward by the full line height
-      // plus spacing.
-      float currentY = anchor[1] + totalHeight * 0.5f;
-      for (size_t i = 0; i < lines.size(); ++i) {
-        CanvasTextStyle style;
-        style.fontFamily = lines[i].fontFamily;
-        style.fontSize = worldFontSizes[i];
-        style.ascent = ascentsWorld[i];
-        style.descent = descentsWorld[i];
-        style.lineHeight = lineHeightsWorld[i];
-        style.extraLineSpacing = lineSpacingWorld;
-        style.color = {0.0f, 0.0f, 0.0f, 1.0f};
-        style.outlineColor = {1.0f, 1.0f, 1.0f, 1.0f};
-        style.outlineWidth = pxToWorld * 0.5f;
-        style.hAlign = CanvasTextStyle::HorizontalAlign::Center;
-        style.vAlign = CanvasTextStyle::VerticalAlign::Baseline;
-        float baseline = currentY - style.ascent;
-        if (ShouldTraceLabelOrder()) {
-          std::ostringstream trace;
-          trace << "[label-capture] fixture=" << uuid << " source="
-                << labelSourceKey << " text=\"" << lines[i].text << "\" x="
-                << anchor[0] << " baseline=" << baseline
-                << " size=" << style.fontSize << " vAlign=Baseline";
-          Logger::Instance().Log(trace.str());
-        }
-        RecordText(anchor[0], baseline, lines[i].text, style);
-        if (i + 1 < lines.size())
-          currentY -= lineHeightsWorld[i] + lineSpacingWorld;
-      }
-    }
-
-    NVGcolor textColor =
-        m_darkMode ? nvgRGBAf(1.f, 1.f, 1.f, 1.f)
-                   : nvgRGBAf(0.f, 0.f, 0.f, 1.f);
-    NVGcolor outlineColor =
-        m_darkMode ? nvgRGBAf(0.f, 0.f, 0.f, 1.f)
-                   : nvgRGBAf(1.f, 1.f, 1.f, 1.f);
-    DrawLabelLines2D(m_vg, lines, x, y, textColor, outlineColor, true);
-  }
+  m_labelRenderSystem->DrawAllFixtureLabels(width, height, view, zoom);
 }
 
 void Viewer3DController::DrawTrussLabels(int width, int height) {
-
-  ConfigManager &cfg = ConfigManager::Get();
-
-  double model[16];
-  double proj[16];
-  int viewport[4];
-  glGetDoublev(GL_MODELVIEW_MATRIX, model);
-  glGetDoublev(GL_PROJECTION_MATRIX, proj);
-  glGetIntegerv(GL_VIEWPORT, viewport);
-
-  const auto hiddenLayers = SnapshotHiddenLayers(cfg);
-  const CullingSettings culling = GetCullingSettings3D(cfg);
-  const float minLabelPixels = culling.minPixels3D;
-  const bool useLabelOptimizations =
-      cfg.GetFloat("label_optimizations_enabled") >= 0.5f;
-  int labelsDrawn = 0;
-  const int maxLabels = GetLabelLimit(cfg, "label_max_trusses");
-  const auto &trusses = SceneDataManager::Instance().GetTrusses();
-  ViewFrustumSnapshot frustum{};
-  std::copy(std::begin(viewport), std::end(viewport), std::begin(frustum.viewport));
-  std::copy(std::begin(model), std::end(model), std::begin(frustum.model));
-  std::copy(std::begin(proj), std::end(proj), std::begin(frustum.projection));
-  const VisibleSet &visibleSet =
-      GetVisibleSet(frustum, hiddenLayers, culling.enabled, minLabelPixels);
-  for (const auto &uuid : visibleSet.trussUuids) {
-    auto trussIt = trusses.find(uuid);
-    if (trussIt == trusses.end())
-      continue;
-    const auto &t = trussIt->second;
-    if (uuid != m_highlightUuid)
-      continue;
-    if (useLabelOptimizations && maxLabels > 0 && labelsDrawn >= maxLabels)
-      break;
-
-    double sx, sy, sz;
-    auto bit = m_trussBounds.find(uuid);
-    if (useLabelOptimizations && culling.enabled && bit != m_trussBounds.end()) {
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max, height,
-                                      model, proj, viewport, rect,
-                                      anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, width, height, minLabelPixels)) {
-        continue;
-      }
-    }
-    if (bit != m_trussBounds.end()) {
-      const BoundingBox &bb = bit->second;
-      double wx = (bb.min[0] + bb.max[0]) * 0.5;
-      double wy = (bb.min[1] + bb.max[1]) * 0.5;
-      double wz = (bb.min[2] + bb.max[2]) * 0.5;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    } else {
-      double wx = t.transform.o[0] * RENDER_SCALE;
-      double wy = t.transform.o[1] * RENDER_SCALE;
-      double wz = t.transform.o[2] * RENDER_SCALE;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    }
-
-    int x = static_cast<int>(sx);
-    int y = height - static_cast<int>(sy);
-    wxString label =
-        t.name.empty() ? wxString::FromUTF8(uuid) : wxString::FromUTF8(t.name);
-    float baseHeight = t.transform.o[2] - t.heightMm * 0.5f;
-    std::string hStr = FormatMeters(baseHeight);
-    label += wxString::Format("\nh = %s m", hStr.c_str());
-
-    auto utf8 = label.ToUTF8();
-    DrawText2D(m_vg, m_font, std::string(utf8.data(), utf8.length()), x, y);
-    ++labelsDrawn;
-  }
+  m_labelRenderSystem->DrawTrussLabels(width, height);
 }
 
 void Viewer3DController::DrawSceneObjectLabels(int width, int height) {
-
-  ConfigManager &cfg = ConfigManager::Get();
-
-  double model[16];
-  double proj[16];
-  int viewport[4];
-  glGetDoublev(GL_MODELVIEW_MATRIX, model);
-  glGetDoublev(GL_PROJECTION_MATRIX, proj);
-  glGetIntegerv(GL_VIEWPORT, viewport);
-
-  const auto hiddenLayers = SnapshotHiddenLayers(cfg);
-  const CullingSettings culling = GetCullingSettings3D(cfg);
-  const float minLabelPixels = culling.minPixels3D;
-  const bool useLabelOptimizations =
-      cfg.GetFloat("label_optimizations_enabled") >= 0.5f;
-  int labelsDrawn = 0;
-  const int maxLabels = GetLabelLimit(cfg, "label_max_objects");
-  const auto &objs = SceneDataManager::Instance().GetSceneObjects();
-  ViewFrustumSnapshot frustum{};
-  std::copy(std::begin(viewport), std::end(viewport), std::begin(frustum.viewport));
-  std::copy(std::begin(model), std::end(model), std::begin(frustum.model));
-  std::copy(std::begin(proj), std::end(proj), std::begin(frustum.projection));
-  const VisibleSet &visibleSet =
-      GetVisibleSet(frustum, hiddenLayers, culling.enabled, minLabelPixels);
-  for (const auto &uuid : visibleSet.objectUuids) {
-    auto objectIt = objs.find(uuid);
-    if (objectIt == objs.end())
-      continue;
-    const auto &o = objectIt->second;
-    if (uuid != m_highlightUuid)
-      continue;
-    if (useLabelOptimizations && maxLabels > 0 && labelsDrawn >= maxLabels)
-      break;
-
-    double sx, sy, sz;
-    auto bit = m_objectBounds.find(uuid);
-    if (useLabelOptimizations && culling.enabled && bit != m_objectBounds.end()) {
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max, height,
-                                      model, proj, viewport, rect,
-                                      anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, width, height, minLabelPixels)) {
-        continue;
-      }
-    }
-    if (bit != m_objectBounds.end()) {
-      const BoundingBox &bb = bit->second;
-      double wx = (bb.min[0] + bb.max[0]) * 0.5;
-      double wy = (bb.min[1] + bb.max[1]) * 0.5;
-      double wz = (bb.min[2] + bb.max[2]) * 0.5;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    } else {
-      double wx = o.transform.o[0] * RENDER_SCALE;
-      double wy = o.transform.o[1] * RENDER_SCALE;
-      double wz = o.transform.o[2] * RENDER_SCALE;
-      if (gluProject(wx, wy, wz, model, proj, viewport, &sx, &sy, &sz) !=
-          GL_TRUE)
-        continue;
-    }
-
-    int x = static_cast<int>(sx);
-    int y = height - static_cast<int>(sy);
-    wxString label =
-        o.name.empty() ? wxString::FromUTF8(uuid) : wxString::FromUTF8(o.name);
-
-    auto utf8 = label.ToUTF8();
-    DrawText2D(m_vg, m_font, std::string(utf8.data(), utf8.length()), x, y);
-    ++labelsDrawn;
-  }
+  m_labelRenderSystem->DrawSceneObjectLabels(width, height);
 }
 
 void Viewer3DController::SetLayerColor(const std::string &layer,
